// types/proto/file.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.7.7
//   protoc               v6.32.0
// source: proto/file.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from '@nestjs/microservices';
import { Observable } from 'rxjs';

export const protobufPackage = 'file';

export interface UploadFileRequest {
  fileName: string;
  /** or base64 string if using JSON transport */
  fileData: Uint8Array;
}

export interface UploadFileResponse {
  fileUrl: string;
}

export interface GetSignedUrlRequest {
  fileName: string;
  expiresIn: number;
}

export interface GetSignedUrlResponse {
  signedUrl: string;
}

export const FILE_PACKAGE_NAME = 'file';

export interface FileServiceClient {
  /** Upload a file and get back its URL */

  uploadFile(request: UploadFileRequest): Observable<UploadFileResponse>;

  getSignedUrl(request: GetSignedUrlRequest): Observable<GetSignedUrlResponse>;
}

export interface FileServiceController {
  /** Upload a file and get back its URL */

  uploadFile(
    request: UploadFileRequest
  ):
    | Promise<UploadFileResponse>
    | Observable<UploadFileResponse>
    | UploadFileResponse;

  getSignedUrl(
    request: GetSignedUrlRequest
  ):
    | Promise<GetSignedUrlResponse>
    | Observable<GetSignedUrlResponse>
    | GetSignedUrlResponse;
}

export function FileServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ['uploadFile', 'getSignedUrl'];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(
        constructor.prototype,
        method
      );
      GrpcMethod('FileService', method)(
        constructor.prototype[method],
        method,
        descriptor
      );
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(
        constructor.prototype,
        method
      );
      GrpcStreamMethod('FileService', method)(
        constructor.prototype[method],
        method,
        descriptor
      );
    }
  };
}

export const FILE_SERVICE_NAME = 'FileService';